requirements.txt
langchain
openai
faiss-cpu
pymupdf
pdfplumber
streamlit
requests
python-dotenv
langchain-community
langchain-openai
tiktoken
PyGithub


